htdocs: Update Bootstrap to v4.1.3
As part of the switch, we need to add a new dependency, popper.js, which
is required by Bootstrap 4. We also need to update the stylesheet for
Selectize [1] so that we use Bootstrap 4 styling where possible.

[1] https://github.com/papakay/selectize-bootstrap-4-style/

Signed-off-by: Stephen Finucane <stephen@that.guru>
---
This should be squashed into the preceding patch as that's useless
without this one.