templates: Upgrade to Bootstrap 4
Bootstrap 4 has been around for a while and offers some pretty neat
features that we can take advantage of in the near future: it's about
time we made the jump.

There are a couple of things required here, meaning this change is
probably bigger than we'd like. That being said, there is a still a lot
more to be done, which we'll be doing in a series of follow-up patches.

htdocs/css/style.css

- various overrides to required for fixed headers and footers and to
  preserve styling between Bootstrap 3 and Bootstrap 4
- remove 'pw-list' class as it was unnecessary
- remove 'navbar-subbrand' as 'navbar-brand' is useful enough

patchwork/filters.py

- the form generated for the 'archived' filter is reworked

patchwork/templates/patchwork/about.html
patchwork/templates/patchwork/partials/download-buttons.html
patchwork/templates/patchwork/partials/filters.html
patchwork/templates/patchwork/partials/patch-list.html
patchwork/templates/patchwork/projects.html
patchwork/templates/patchwork/submission.html

- rework all elements to replace removed Bootstrap 3 elements with their
  Bootstrap 4 equivalents

templates/base.html

- update reference to bootstrap4 libs and include reference to popper.js
- the navbar have been reworked using new bootstrap styles
- the footer has been reworked to be sticky; this also includes some
  changes around which parts of the footer are clickable

In addition, as part of the switch, we need to add a new dependency,
popper.js, which is required by Bootstrap 4. We also need to update the
stylesheet for Selectize [1] so that we use Bootstrap 4 styling where
possible.

For a detailed overview of the changes, refer to the Bootstrap 4
migration guide [2].

[1] https://github.com/papakay/selectize-bootstrap-4-style/
[2] https://getbootstrap.com/docs/4.0/migration/

Signed-off-by: Stephen Finucane <stephen@that.guru>
---
The new CSS/JS files are sent in a separate patch to avoid the whole
thing getting dropped by Mailman.