htdocs: Add and integrate Font Awesome
As part of the switch to Bootstrap 4, we need to decide what to replace
Glyphicons with. While we could continue using these and simply add our
own CSS [1], Font Awesome has certainly gained traction in recent years
and looks like the best option available to us. Let's use it.

We can't actually get rid of Glypicons yet as our build of Bootstrap
relies on them. We can do that when we upgrade to Bootstrap 4 though.

[1] http://blog.creative-tim.com/web-design/bootstrap-glyphicons/

Signed-off-by: Stephen Finucane <stephen@that.guru>